+++ src/anime_parsers_ru/__init__.py
@@ -1,3 +1,4 @@
 from .parser_kodik import KodikParser
 from .parser_aniboom import AniboomParser
 from .parser_jutsu import JutsuParser
+from .parser_shikimori import ShikimoriParser